RobAF-Mark/requirements.txt
kornia==0.7.2
matplotlib==3.8.2
numpy==1.26.4
Pillow==10.0.0
Pillow==10.2.0
scipy==1.12.0
skimage==0.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.1
